+++ turbopack/crates/turbopack-ecmascript-runtime/js/package.json
@@ -8,9 +8,9 @@
     "check": "run-p check:*",
     "check:nodejs": "tsc -p src/nodejs",
     "check:browser-dev-client": "tsc -p src/browser/dev/hmr-client",
-    "check:browser-dev-runtime-base": "tsc -p src/browser/dev/runtime/base",
-    "check:browser-dev-runtime-dom": "tsc -p src/browser/dev/runtime/dom",
-    "check:browser-dev-runtime-edge": "tsc -p src/browser/dev/runtime/edge"
+    "check:browser-runtime-base": "tsc -p src/browser/runtime/base",
+    "check:browser-runtime-dom": "tsc -p src/browser/runtime/dom",
+    "check:browser-runtime-edge": "tsc -p src/browser/runtime/edge"
   },
   "exports": {
     ".": "./src/main.js",
