+++ turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/base/tsconfig.json
@@ -1,8 +1,8 @@
 {
-  "extends": "../../../../tsconfig.base.json",
+  "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
     // environment, we need WebWorker for WebAssembly types
     "lib": ["ESNext", "WebWorker"]
   },
-  "include": ["*.ts"]
+  "include": ["runtime-base.ts", "dummy.ts"]
 }
