+++ turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/dom/runtime-backend-dom.ts
@@ -7,8 +7,8 @@
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
-/// <reference path="../base/runtime-base.ts" />
-/// <reference path="../../../../shared/require-type.d.ts" />
+/// <reference path="../../../browser/runtime/base/runtime-base.ts" />
+/// <reference path="../../../shared/runtime-types.d.ts" />
 
 type ChunkResolver = {
   resolved: boolean;
@@ -19,7 +19,7 @@ type ChunkResolver = {
 
 let BACKEND: RuntimeBackend;
 
-function augmentContext(context: TurbopackDevBaseContext): TurbopackDevContext {
+function augmentContext(context: unknown): unknown {
   return context;
 }
 
@@ -28,7 +28,7 @@ function fetchWebAssembly(wasmChunkPath: ChunkPath) {
 }
 
 async function loadWebAssembly(
-  _source: SourceInfo,
+  _source: unknown,
   wasmChunkPath: ChunkPath,
   importsObj: WebAssembly.Imports
 ): Promise<Exports> {
@@ -40,13 +40,18 @@ async function loadWebAssembly(
 }
 
 async function loadWebAssemblyModule(
-  _source: SourceInfo,
+  _source: unknown,
   wasmChunkPath: ChunkPath
 ): Promise<WebAssembly.Module> {
   const req = fetchWebAssembly(wasmChunkPath);
 
   return await WebAssembly.compileStreaming(req);
 }
+
+/**
+ * Maps chunk paths to the corresponding resolver.
+ */
+const chunkResolvers: Map<ChunkPath, ChunkResolver> = new Map();
 
 (() => {
   BACKEND = {
@@ -81,103 +86,7 @@ async function loadWebAssemblyModule(
     loadChunk(chunkPath, source) {
       return doLoadChunk(chunkPath, source);
     },
-
-    unloadChunk(chunkPath) {
-      deleteResolver(chunkPath);
-
-      const chunkUrl = getChunkRelativeUrl(chunkPath);
-      // TODO(PACK-2140): remove this once all filenames are guaranteed to be escaped.
-      const decodedChunkUrl = decodeURI(chunkUrl);
-
-      if (chunkPath.endsWith(".css")) {
-        const links = document.querySelectorAll(
-          `link[href="${chunkUrl}"],link[href^="${chunkUrl}?"],link[href="${decodedChunkUrl}"],link[href^="${decodedChunkUrl}?"]`
-        );
-        for (const link of Array.from(links)) {
-          link.remove();
-        }
-      } else if (chunkPath.endsWith(".js")) {
-        // Unloading a JS chunk would have no effect, as it lives in the JS
-        // runtime once evaluated.
-        // However, we still want to remove the script tag from the DOM to keep
-        // the HTML somewhat consistent from the user's perspective.
-        const scripts = document.querySelectorAll(
-          `script[src="${chunkUrl}"],script[src^="${chunkUrl}?"],script[src="${decodedChunkUrl}"],script[src^="${decodedChunkUrl}?"]`
-        );
-        for (const script of Array.from(scripts)) {
-          script.remove();
-        }
-      } else {
-        throw new Error(`can't infer type of chunk from path ${chunkPath}`);
-      }
-    },
-
-    reloadChunk(chunkPath) {
-      return new Promise<void>((resolve, reject) => {
-        if (!chunkPath.endsWith(".css")) {
-          reject(new Error("The DOM backend can only reload CSS chunks"));
-          return;
-        }
-
-        const chunkUrl = getChunkRelativeUrl(chunkPath);
-        const decodedChunkUrl = decodeURI(chunkUrl);
-
-        const previousLinks = document.querySelectorAll(
-          `link[rel=stylesheet][href="${chunkUrl}"],link[rel=stylesheet][href^="${chunkUrl}?"],link[rel=stylesheet][href="${decodedChunkUrl}"],link[rel=stylesheet][href^="${decodedChunkUrl}?"]`
-        );
-
-        if (previousLinks.length === 0) {
-          reject(new Error(`No link element found for chunk ${chunkPath}`));
-          return;
-        }
-
-        const link = document.createElement("link");
-        link.rel = "stylesheet";
-
-        if (navigator.userAgent.includes("Firefox")) {
-          // Firefox won't reload CSS files that were previously loaded on the current page,
-          // we need to add a query param to make sure CSS is actually reloaded from the server.
-          //
-          // I believe this is this issue: https://bugzilla.mozilla.org/show_bug.cgi?id=1037506
-          //
-          // Safari has a similar issue, but only if you have a `<link rel=preload ... />` tag
-          // pointing to the same URL as the stylesheet: https://bugs.webkit.org/show_bug.cgi?id=187726
-          link.href = `${chunkUrl}?ts=${Date.now()}`;
-        } else {
-          link.href = chunkUrl;
-        }
-
-        link.onerror = () => {
-          reject();
-        };
-        link.onload = () => {
-          // First load the new CSS, then remove the old ones. This prevents visible
-          // flickering that would happen in-between removing the previous CSS and
-          // loading the new one.
-          for (const previousLink of Array.from(previousLinks))
-            previousLink.remove();
-
-          // CSS chunks do not register themselves, and as such must be marked as
-          // loaded instantly.
-          resolve();
-        };
-
-        // Make sure to insert the new CSS right after the previous one, so that
-        // its precedence is higher.
-        previousLinks[0].parentElement!.insertBefore(
-          link,
-          previousLinks[0].nextSibling
-        );
-      });
-    },
-
-    restart: () => self.location.reload(),
   };
-
-  /**
-   * Maps chunk paths to the corresponding resolver.
-   */
-  const chunkResolvers: Map<ChunkPath, ChunkResolver> = new Map();
 
   function getOrCreateResolver(chunkPath: ChunkPath): ChunkResolver {
     let resolver = chunkResolvers.get(chunkPath);
@@ -200,10 +109,6 @@ async function loadWebAssemblyModule(
       chunkResolvers.set(chunkPath, resolver);
     }
     return resolver;
-  }
-
-  function deleteResolver(chunkPath: ChunkPath) {
-    chunkResolvers.delete(chunkPath);
   }
 
   /**
@@ -299,19 +204,3 @@ async function loadWebAssemblyModule(
     return resolver.promise;
   }
 })();
-
-function _eval({ code, url, map }: EcmascriptModuleEntry): ModuleFactory {
-  code += `\n\n//# sourceURL=${encodeURI(
-    location.origin + CHUNK_BASE_PATH + url
-  )}`;
-  if (map) {
-    code += `\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,${btoa(
-      // btoa doesn't handle nonlatin characters, so escape them as \x sequences
-      // See https://stackoverflow.com/a/26603875
-      unescape(encodeURIComponent(map))
-    )}`;
-  }
-
-  // eslint-disable-next-line no-eval
-  return eval(code);
-}
