+++ turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/dom/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../../tsconfig.base.json",
+  "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
     // environment
     "lib": ["ESNext", "DOM", "WebWorker.ImportScripts"]
