+++ turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/edge/tsconfig.json
@@ -1,8 +1,8 @@
 {
-  "extends": "../../../../tsconfig.base.json",
+  "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
     // environment, we need WebWorker for WebAssembly types
     "lib": ["ESNext", "WebWorker"]
   },
-  "include": ["runtime-base.ts", "dummy.ts"]
+  "include": ["*.ts"]
 }
